backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose==3.3.0
openai==1.108.1
